+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Bug: sensor and fire control installed systems were missign the edit button.
 - Bug: computer cores could be set to invalid sizes.
 - Bug: now ships can have more than 1 computer core.
+- Bug: multiple of the same support system can now be installed.
 
 ## [0.2.1] - 2026-02-04